[ui] Add old quit keybind
Between v0.1.3 and v0.1.7, the change was made to change the default
'Quit' keybind was changed from 'q' to 'Shift+q'. Additionally the quit
confirmation no longer displays. The help menu has been updated and
previous behavior restored in a way that does not conflict with the
updated quit behavior.

Signed-off-by: Jack Champagne <jackchampagne.r@gmail.com>